requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.0

llama-index>=0.10.30
llama-index-embeddings-huggingface>=0.1.4
llama-index-retrievers-bm25>=0.1.3

sentence-transformers>=2.6.1
pypdf>=4.2.0

langchain>=0.1.20
langgraph>=0.0.40
langchain-groq>=0.1.4
